fixed #82 and #70
